Move RenderSprites and RenderSimple into Mods.RA.
