+++ OpenRA.Mods.Cnc/Widgets/ProductionPaletteWidget.cs
@@ -16,6 +16,7 @@ using OpenRA.Graphics;
 using OpenRA.Mods.RA;
 using OpenRA.Mods.RA.Buildings;
 using OpenRA.Mods.RA.Orders;
+using OpenRA.Mods.RA.Render;
 using OpenRA.Network;
 using OpenRA.Traits;
 using OpenRA.Widgets;
