+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -520,6 +520,8 @@
     <Compile Include="Widgets\Logic\InstallMusicLogic.cs" />
     <Compile Include="Render\WithActiveAnimation.cs" />
     <Compile Include="SupportPowers\SpawnActorPower.cs" />
+    <Compile Include="Render\RenderSimple.cs" />
+    <Compile Include="Render\RenderSprites.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\OpenRA.Game\OpenRA.Game.csproj">
