+++ OpenRA.Mods.RA/Render/RenderSprites.cs
@@ -12,9 +12,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Graphics;
+using OpenRA.Traits;
 using OpenRA.Primitives;
 
-namespace OpenRA.Traits
+namespace OpenRA.Mods.RA.Render
 {
 	public class RenderSpritesInfo : ITraitInfo
 	{
