+++ OpenRA.Mods.RA/SupportPowers/NukePower.cs
@@ -12,6 +12,7 @@ using System;
 using OpenRA.Effects;
 using OpenRA.Mods.RA.Activities;
 using OpenRA.Mods.RA.Effects;
+using OpenRA.Mods.RA.Render;
 using OpenRA.Primitives;
 using OpenRA.Traits;
