+++ OpenRA.Mods.RA/Widgets/BuildPaletteWidget.cs
@@ -15,6 +15,7 @@ using System.Linq;
 using OpenRA.Graphics;
 using OpenRA.Mods.RA.Buildings;
 using OpenRA.Mods.RA.Orders;
+using OpenRA.Mods.RA.Render;
 using OpenRA.Network;
 using OpenRA.Primitives;
 using OpenRA.Traits;
