+++ OpenRA.Mods.RA/Widgets/ObserverProductionIconsWidget.cs
@@ -13,6 +13,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using OpenRA.Graphics;
+using OpenRA.Mods.RA.Render;
 using OpenRA.Traits;
 using OpenRA.Widgets;
